hotfix: safe area insets implemented - minor refactoring on tab navigations
