+++ components/navigation/index.ts
@@ -1,0 +1,2 @@
+export { default as TabNavigator } from "./TabNavigator";
+export { default as SettingsStack } from "./SettingsStack";
